compiler-rt: Fix __floatunsitf signature
The function transforms an unsigned integer into a f128.

Closes #8996

+++ lib/std/special/compiler_rt/floatunsitf_test.zig
@@ -5,7 +5,7 @@
 // and substantial portions of the software.
 const __floatunsitf = @import("floatunsitf.zig").__floatunsitf;
 
-fn test__floatunsitf(a: u64, expected_hi: u64, expected_lo: u64) !void {
+fn test__floatunsitf(a: u32, expected_hi: u64, expected_lo: u64) !void {
     const x = __floatunsitf(a);
 
     const x_repr = @bitCast(u128, x);
